requirements.txt
questsFlask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
requests